+++ config/routes.rb
@@ -2,5 +2,7 @@
 
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  resources :turtles, except: %i[new edit]
+  namespace :v1 do 
+    resources :turtles, except: %i[new edit]
+  end
 end
